requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-filter==24.3
django-jazzmin==3.0.0
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.9
magic-filter==1.0.12
multidict==6.1.0
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyTelegramBotAPI==4.22.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
yarl==1.11.1
